Restructure the application to be Onion
