+++ nestjs8.2.5/src/infrastructure/application/default-operation-response.resolver.ts
@@ -1,0 +1,8 @@
+import { OperationResponseDto } from "@services/application/dto";
+import { IOperationResponseResolver } from "@services/application/operation-response-resolver";
+
+export class DefaultOperationResponseResolver implements IOperationResponseResolver {
+  resolve(value: number): OperationResponseDto {
+    return { value };
+  }
+}
