+++ nestjs8.2.5/src/infrastructure/application/operations/subtraction.operation.ts
@@ -1,0 +1,12 @@
+import { IOperation } from '@services/application/operation';
+import { OperationInputDto } from '@services/application/dto';
+
+export class SubtractionOperation implements IOperation {
+  public produce(inputDto: OperationInputDto): number {
+    if (!inputDto) {
+      throw new Error(`Argument must be provided`);
+    }
+
+    return inputDto.leftOperand - inputDto.rightOperand;
+  }
+}
