--- nestjs8.2.5/src/operations/implementations/addition.operation.ts
@@ -1,8 +1,0 @@
-import { IOperation } from '../operation.interface';
-import { OperationInputDto } from '../dto/operation-input.dto';
-
-export class AdditionOperation implements IOperation {
-  public produce(inputDto: OperationInputDto): number {
-    return inputDto.leftOperand + inputDto.rightOperand;
-  }
-}
