--- nestjs8.2.5/src/operations/implementations/default-operation-response.resolver.ts
@@ -1,10 +1,0 @@
-import { OperationResponseDto } from '../dto/operation-response.dto';
-import { IOperationResponseResolver } from '../operation-response-resolver.interface';
-
-export class DefaultOperationResponseResolver
-  implements IOperationResponseResolver
-{
-  resolve(value: number): OperationResponseDto {
-    return { value };
-  }
-}
