--- nestjs8.2.5/src/operations/implementations/division.operation.ts
@@ -1,11 +1,0 @@
-import { IOperation } from '../operation.interface';
-import { OperationInputDto } from '../dto/operation-input.dto';
-
-export class DivisionOperation implements IOperation {
-  public produce(inputDto: OperationInputDto): number {
-    if (inputDto.rightOperand == 0) {
-      throw new Error('Right operand must not equal to 0');
-    }
-    return inputDto.leftOperand / inputDto.rightOperand;
-  }
-}
