--- nestjs8.2.5/src/operations/implementations/subtraction.operation.ts
@@ -1,8 +1,0 @@
-import { IOperation } from '../operation.interface';
-import { OperationInputDto } from '../dto/operation-input.dto';
-
-export class SubtractionOperation implements IOperation {
-  public produce(inputDto: OperationInputDto): number {
-    return inputDto.leftOperand - inputDto.rightOperand;
-  }
-}
