--- nestjs8.2.5/src/operations/operation-response-resolver.interface.ts
@@ -1,5 +1,0 @@
-import { OperationResponseDto } from './dto/operation-response.dto';
-
-export interface IOperationResponseResolver {
-  resolve(value: number): OperationResponseDto;
-}
