--- nestjs8.2.5/src/operations/operation.interface.ts
@@ -1,5 +1,0 @@
-import { OperationInputDto } from './dto/operation-input.dto';
-
-export interface IOperation {
-  produce(inputDto: OperationInputDto): number;
-}
