--- nestjs8.2.5/src/operations/operations.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { OperationsService } from './operations.service';
-
-describe('OperationsService', () => {
-  let service: OperationsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [OperationsService],
-    }).compile();
-
-    service = module.get<OperationsService>(OperationsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
