--- nestjs8.2.5/src/operations/providers/addition-operation.provider.ts
@@ -1,8 +1,0 @@
-import { Provider } from '@nestjs/common';
-import { IOperation } from '../operation.interface';
-import { AdditionOperation } from '../implementations/addition.operation';
-
-export const AdditionOperationProvider: Provider<IOperation> = {
-  provide: 'OPERATION+',
-  useClass: AdditionOperation,
-};
