--- nestjs8.2.5/src/operations/providers/default-operation-response-resolver.provider.ts
@@ -1,9 +1,0 @@
-import { Provider } from '@nestjs/common';
-import { IOperationResponseResolver } from '../operation-response-resolver.interface';
-import { DefaultOperationResponseResolver } from '../implementations/default-operation-response.resolver';
-
-export const DefaultOperationResponseResolverProvider: Provider<IOperationResponseResolver> =
-  {
-    provide: 'RESPONSE',
-    useClass: DefaultOperationResponseResolver,
-  };
