--- nestjs8.2.5/src/operations/providers/division-operation.provider.ts
@@ -1,8 +1,0 @@
-import { Provider } from '@nestjs/common';
-import { IOperation } from '../operation.interface';
-import { DivisionOperation } from '../implementations/division.operation';
-
-export const DivisionOperationProvider: Provider<IOperation> = {
-  provide: 'OPERATION/',
-  useClass: DivisionOperation,
-};
