--- nestjs8.2.5/src/operations/providers/multiplication-operation.provider.ts
@@ -1,8 +1,0 @@
-import { Provider } from '@nestjs/common';
-import { IOperation } from '../operation.interface';
-import { MultiplicationOperation } from '../implementations/multiplication.operation';
-
-export const MultiplicationOperationProvider: Provider<IOperation> = {
-  provide: 'OPERATION*',
-  useClass: MultiplicationOperation,
-};
