--- nestjs8.2.5/src/operations/providers/subtraction-operation.provider.ts
@@ -1,8 +1,0 @@
-import { Provider } from '@nestjs/common';
-import { IOperation } from '../operation.interface';
-import { SubtractionOperation } from '../implementations/subtraction.operation';
-
-export const SubtractionOperationProvider: Provider<IOperation> = {
-  provide: 'OPERATION-',
-  useClass: SubtractionOperation,
-};
