+++ nestjs8.2.5/src/providers/addition-operation.provider.ts
@@ -1,0 +1,8 @@
+import { Provider } from '@nestjs/common';
+import { IOperation } from '@services/application/operation';
+import { AdditionOperation } from '@infrastructure/application/operations/addition.operation';
+
+export const AdditionOperationProvider: Provider<IOperation> = {
+  provide: 'OPERATION+',
+  useClass: AdditionOperation,
+};
