+++ nestjs8.2.5/src/providers/multiplication-operation.provider.ts
@@ -1,0 +1,8 @@
+import { Provider } from '@nestjs/common';
+import { IOperation } from '@services/application/operation';
+import { MultiplicationOperation } from '@infrastructure/application/operations/multiplication.operation';
+
+export const MultiplicationOperationProvider: Provider<IOperation> = {
+  provide: 'OPERATION*',
+  useClass: MultiplicationOperation,
+};
