+++ nestjs8.2.5/src/services/application/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './operation-input.dto';
+export * from './operation-request.dto';
+export * from './operation-response.dto';
