+++ nestjs8.2.5/src/services/application/module.ts
@@ -1,0 +1,5 @@
+import { Module } from '@nestjs/common';
+
+@Module({
+})
+export class ApplicationServicesModule {}
