+++ nestjs8.2.5/src/services/application/operation-response-resolver.ts
@@ -1,0 +1,6 @@
+import { OperationResponseDto } from "./dto";
+
+export interface IOperationResponseResolver {
+    resolve(value: number): OperationResponseDto;
+}
+
