+++ nestjs8.2.5/src/services/application/operation.ts
@@ -1,0 +1,5 @@
+import { OperationInputDto } from "./dto";
+
+export interface IOperation {
+    produce(inputDto: OperationInputDto): number;
+}
