+++ nestjs8.2.5/src/services/application/operations.service.ts
@@ -1,0 +1,5 @@
+import { OperationRequestDto, OperationResponseDto } from "./dto";
+
+export interface IOperationsService {
+    handle(requestDto: OperationRequestDto): Promise<OperationResponseDto>;
+}
