+++ nestjs8.2.5/src/spec/infrastructure/application/operations/addition.operation.spec.ts
@@ -1,0 +1,34 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { IOperation } from '@services/application/operation';
+import { OperationInputDto } from '@services/application/dto';
+import { AdditionOperationProvider } from '@providers/addition-operation.provider';
+import { ApplicationInfrastructureModule } from '@infrastructure/application/module';
+import { AdditionOperation } from '@infrastructure/application/operations/addition.operation';
+
+describe(AdditionOperation, () => {
+  let operation: IOperation;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [AdditionOperationProvider],
+      imports: [ApplicationInfrastructureModule],
+    }).compile();
+
+    operation = module.get<IOperation>('OPERATION+');
+  });
+
+  it('should be defined', () => {
+    expect(operation).toBeDefined();
+  });
+
+  it('should recieve 3 and 5 and return 8', () => {
+    const dto: OperationInputDto = {
+      leftOperand: 5,
+      rightOperand: 3,
+    };
+    const result: number = operation.produce(dto);
+
+    expect(result).not.toEqual(null);
+    expect(result).toBe(8);
+  });
+});
