+++ nestjs8.2.5/tsconfig.json
@@ -10,6 +10,12 @@
     "sourceMap": true,
     "outDir": "./dist",
     "baseUrl": "./",
+    "paths": {
+      "@services/*": ["src/services/*"],
+      "@providers/*": ["src/providers/*"],
+      "@controllers/*": ["src/controllers/*"],
+      "@infrastructure/*": ["src/infrastructure/*"]
+    },
     "incremental": true,
     "skipLibCheck": true,
     "strictNullChecks": false,
